backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
httpx
pydantic
pydantic-settings
python-multipart
jose
passlib
bcrypt
phonenumbers
dnspython
aiohttp
